create a default error table
