+++ components/pages/submission-system/program-clinical-submission/PageContent.tsx
@@ -51,11 +51,12 @@ import { useState, useEffect, useMemo, ComponentProps, createRef } from 'react';
 
 import { useClinicalSubmissionQuery } from '.';
 import { containerStyle } from '../common';
-import ErrorNotification, {
+import ErrorNotification from './ErrorNotification';
+import {
   ErrorNotificationDefaultColumns,
   errorNotificationTableProps,
   getDefaultColumns,
-} from '../ErrorNotification';
+} from './ErrorNotification/ErrorNotificationDefaultTable';
 import { SchemaInvalidSubmissionNotification } from '../SchemaInvalidSubmissionNotification';
 import {
   SubmissionSystemLockedNotification,
